move subscription dashboard queries to dashboard project
