+++ src/Beckett.Dashboard/Subscriptions/Resume/ResumeEndpoint.cs
@@ -13,7 +13,7 @@ public static class ResumeEndpoint
         CancellationToken cancellationToken
     )
     {
-        await database.Execute(new Beckett.Subscriptions.Queries.ResumeSubscription(groupName, name, options), cancellationToken);
+        await database.Execute(new ResumeQuery(groupName, name, options), cancellationToken);
 
         context.Response.Headers.Append("HX-Refresh", new StringValues("true"));
